Update README.md
Hi!

I just started searching for a mod manager since I normally do these install manually.

It worked fine for me because I normally used a few mods and doing “by hand” was fine. But now I want to experiment with a list much bigger than I normally use, so I want to try a mod manager.

On Linux, I'm quite savvy, knowing how to search for answers for any type of distribution and then using logic/intuition to convert for my specific necessity and Distro that I'm currently on.

For gaming, it's Nobara [https://nobaraproject.org/]. For fun, Chimera Linux [https://chimera-linux.org/].

So, reading your README and trying to build your tool was relatively straightforward. Found the correct way to install Rust [https://developer.fedoraproject.org/tech/languages/rust/rust-installation.html] and then reading the errors, I could search for the missing packages the building was requiring.

Since many Linux gamers just know to follow the “recipe” list of commands, I decided to contribute with an example of how to build in one of the Distros these Linux Gamers use. To simply follow and build without leading with problems!

I did a recursive test to know the exact leaf packages that are needed when you try to build to put a dnf install command as short as it was possible. Doing my research, seems to me that xdg-utils will already be installed if the user is using any of the WM available, so I didn't include in the list.

So, know I will begin to try to use your dmodman tool! I built, revised the steps for building with Nobara and wrote all of this before testing the app itself. : )